author is relatively in the same spot, vertically
this is done by adding a flex-grow value to each of the text child elements in the .text class (these are the flex items). i had to add a div around the quote text so that i can control where the quote text actually is. that div around the quote text, called .quote-container, is also a flex-container. this is the easiest way for me to control this